Made signINGooglehelper file -LE
