+++ UNCUTApp/UNCUTApp/Authentication/AuthView.swift
@@ -8,28 +8,13 @@
 import SwiftUI
 import GoogleSignIn
 import GoogleSignInSwift
-import FirebaseAuth
-
-struct GoogleSIgnINResultModel {
-    let idToken: String
-    let accessToken: String
-}
 
 @MainActor
 final class AuthViewModel: ObservableObject {
     
     func signInGoogle() async throws{
-        guard let topVC = Utilities.shared.topViewController() else {
-            throw URLError(.cannotFindHost)
-        }
-        let gidSignInResult = try await GIDSignIn.sharedInstance.signIn(withPresenting: topVC)
-        
-        
-        guard let idToken: String = gidSignInResult.user.idToken?.tokenString else{
-            throw URLError(.badServerResponse)
-        }
-        let accessToken: String = gidSignInResult.user.accessToken.tokenString
-        let tokens = GoogleSIgnINResultModel(idToken: idToken, accessToken: accessToken)
+        let helper = GoogleSignInHelper()
+        let tokens = try await helper.signIn()
         try await AuthManager.shared.signInWithGoogle(tokens: tokens)
         
     }
@@ -64,17 +49,6 @@ struct AuthView: View {
                         }
                         }
                 })
-                
-                    
-//                NavigationLink(destination: Text("Sign in with Google")) {
-//                    Text("Sign in with Google")
-//                        .font(.headline)
-//                        .foregroundColor(.white)
-//                        .frame(height: 55)
-//                        .frame(maxWidth: .infinity)
-//                        .background(Color.red) // You can change the color
-//                        .cornerRadius(10)
-//                }
             }
             .padding()
             .navigationTitle("Sign In")
